DSM-0063   apply difficulty levels to the dungeon    - number of mobs    - number of items / drop tables    - item levels / upgrade failures / cursed status

+++ src/main/java/com/shatteredpixel/pixeldungeonunleashed/levels/features/HighGrass.java
@@ -47,7 +47,42 @@ import com.watabou.utils.Random;
 public class HighGrass {
 
 	public static void trample( Level level, int pos, Char ch ) {
-		
+		int naturalDrop = 25;  // used for seed and dew drops
+		int foodDrop = 50;     // used for food, id scrolls and healing potions
+		int unusualDrop = 100; // used for scrolls, potions and rings
+
+		switch (Dungeon.difficultyLevel) {
+			case Dungeon.DIFF_TUTOR:
+				naturalDrop = 16;
+				foodDrop = 22;
+				unusualDrop = 80;
+				break;
+			case Dungeon.DIFF_EASY:
+				naturalDrop = 16;
+				foodDrop = 22;
+				unusualDrop = 80;
+				break;
+			case Dungeon.DIFF_NORM:
+				naturalDrop = 18;
+				foodDrop = 50;
+				unusualDrop = 125;
+				break;
+			case Dungeon.DIFF_HARD:
+				naturalDrop = 24;
+				foodDrop = 600;
+				unusualDrop = 800;
+				break;
+			case Dungeon.DIFF_NTMARE:
+				naturalDrop = 30;
+				foodDrop = 1000;
+				unusualDrop = 1200;
+				break;
+			default:
+				naturalDrop = 18;
+				foodDrop = 50;
+				unusualDrop = 125;
+				break;
+		}
 		Level.set( pos, Terrain.GRASS );
 		GameScene.updateMap( pos );
 
@@ -68,7 +103,7 @@ public class HighGrass {
 
 			if (naturalismLevel >= 0) {
 				// Seed, scales from 1/16 to 1/4
-				if (Random.Int(16 - ((int) (naturalismLevel * 3))) == 0) {
+				if (Random.Int(naturalDrop - ((int) (naturalismLevel * 3))) == 0) {
 					Item seed = Generator.random(Generator.Category.SEED);
 
 					if (seed instanceof BlandfruitBush.Seed) {
@@ -81,54 +116,54 @@ public class HighGrass {
 				}
 
 				// Dew, scales from 1/6 to 1/3
-				if (Random.Int(24 - naturalismLevel*3) <= 3) {
+				if (Random.Int((naturalDrop + 8) - naturalismLevel*3) <= 3) {
 					level.drop(new Dewdrop(), pos).sprite.drop();
 				}
 			}
 
-			if ((Random.Int(50) == 0) && (!Dungeon.limitedDrops.seedBag.dropped())) {
+			if ((Dungeon.difficultyLevel <= Dungeon.DIFF_EASY) && (Random.Int(60) == 0) && (!Dungeon.limitedDrops.seedBag.dropped())) {
 				level.drop(new SeedPouch(), pos).sprite.drop();
 				Dungeon.limitedDrops.seedBag.count++;
 			}
 		}
 
 		if (!Dungeon.isChallenged( Challenges.NO_FOOD )) {
-			if (Random.Int(25) == 0) {
+			if (Random.Int(foodDrop) == 0) {
 				Item food = Generator.random(Generator.Category.FOOD);
 				level.drop(food, pos).sprite.drop();
 			}
 		}
 
 		if (!Dungeon.isChallenged( Challenges.NO_HEALING )) {
-			if (Random.Int(22) == 0) {
+			if (Random.Int(foodDrop) == 0) {
 				level.drop(new PotionOfHealing().identify(), pos).sprite.drop();
 			}
 		}
 
-		if (Random.Int(20) == 0) {
+		if ((Dungeon.difficultyLevel <= Dungeon.DIFF_NORM) && Random.Int(20) == 0) {
 			level.drop(new Dart(Random.Int(3)+1), pos).sprite.drop();
 		}
 
 		if (!Dungeon.isChallenged( Challenges.NO_SCROLLS )) {
-			if ((Random.Int(50) == 0) && (!Dungeon.limitedDrops.scrollBag.dropped())) {
+			if ((Dungeon.difficultyLevel <= Dungeon.DIFF_EASY) && (Random.Int(60) == 0) && (!Dungeon.limitedDrops.scrollBag.dropped())) {
 				level.drop(new ScrollHolder(), pos).sprite.drop();
 				Dungeon.limitedDrops.scrollBag.count++;
 			}
 
-			if (Random.Int(25) == 0) {
+			if (Random.Int(foodDrop) == 0) {
 				level.drop(new ScrollOfIdentify().identify(), pos).sprite.drop();
 			}
 
-			if (Random.Int(80) == 0) {
+			if (Random.Int(unusualDrop) == 0) {
 				level.drop(Generator.random(Generator.Category.SCROLL), pos).sprite.drop();
 			}
 		}
 
-		if (Random.Int(80) == 0) {
+		if (Random.Int(unusualDrop) == 0) {
 			level.drop(Generator.random(Generator.Category.POTION), pos).sprite.drop();
 		}
 
-		if (Random.Int(100) == 0) {
+		if (Random.Int(unusualDrop + 25) == 0) {
 			level.drop(Generator.random(Generator.Category.RING), pos).sprite.drop();
 		}
